refactor: initialization 패키지 구조 변경 및 위치 수정
