+++ src/main/java/com/cliptripbe/global/config/initialization/initializer/DataInitializer.java
@@ -1,16 +1,14 @@
-package com.cliptripbe.infrastructure.inital;
+package com.cliptripbe.global.config.initialization.initializer;
 
-import static com.cliptripbe.infrastructure.inital.type.DefaultData.ACCOMMODATION_SEOUL;
-import static com.cliptripbe.infrastructure.inital.type.DefaultData.BUSAN_ACCESSIBLE_TOURISM;
-import static com.cliptripbe.infrastructure.inital.type.DefaultData.INCHEON_ACCESSIBLE_TOURISM;
-import static com.cliptripbe.infrastructure.inital.type.DefaultData.SOKCHO_OPEN_TOURISM;
-import static com.cliptripbe.infrastructure.inital.type.DefaultData.STORAGE_SEOUL;
+import static com.cliptripbe.global.config.initialization.type.DefaultData.ACCOMMODATION_SEOUL;
+import static com.cliptripbe.global.config.initialization.type.DefaultData.BUSAN_ACCESSIBLE_TOURISM;
+import static com.cliptripbe.global.config.initialization.type.DefaultData.INCHEON_ACCESSIBLE_TOURISM;
+import static com.cliptripbe.global.config.initialization.type.DefaultData.SOKCHO_OPEN_TOURISM;
+import static com.cliptripbe.global.config.initialization.type.DefaultData.STORAGE_SEOUL;
 
 import com.cliptripbe.feature.place.domain.entity.Place;
 import com.cliptripbe.feature.place.infrastructure.PlaceRepository;
-import com.cliptripbe.infrastructure.inital.initaler.BookmarkInitializer;
-import com.cliptripbe.infrastructure.inital.initaler.PlaceInitializer;
-import com.cliptripbe.infrastructure.inital.type.DefaultData;
+import com.cliptripbe.global.config.initialization.type.DefaultData;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
