+++ src/main/java/com/cliptripbe/global/config/initialization/mapper/PlaceMapper.java
@@ -1,4 +1,4 @@
-package com.cliptripbe.infrastructure.inital.initaler.mapper;
+package com.cliptripbe.global.config.initialization.mapper;
 
 import com.cliptripbe.feature.place.domain.entity.Place;
 import com.cliptripbe.feature.place.domain.type.PlaceType;
